requirements.txt
streamlit
google-generativeai
msal
requests
python-docx
openpyxl
